functions/requirements.txt
pillow
opencv-python
numpy
google-cloud-storage
